Slightly better performance of LinkedList in this case.
Btw the iteration can be replaced by a map )i'm not sure if sorted or not). This would make 
finding a threadInfo of the current priority easy, at the cost of some bookkeeping when
changing priorities. I don't understand why does it rotate the thread infos though?
is it just a adaptative measure? Then a map would really eliminate this overhead.